Moved the old document to "draft" and rewrote a lighter suggestion for  certification standard.
It still lacks complance to CIP-0052, reference to the certification level and where they need to be taken into account.
The text is still too "test" oriented and doesn't easily leave room for formal verification, maybe it can be changed in the text with "verified" instead of "tested".
We'll be able to loop back to certification levels  when we better describe the things the developer (or auditor, or sponsor if we want to add them?) needs to deliver for the evaluation. We could have layers for spec for example, level 3 requiring a formal spec and a formal verification done on those spec and/or on the conformance of source code to formal spec.